split TorConfig unit test into separate subclass tests
