+++ library/controller/kmp-tor-controller-common/src/commonTest/kotlin/io/matthewnelson/kmp/tor/controller/common/config/TorConfig_Option_Time_UnitTest.kt
@@ -1,0 +1,45 @@
+/*
+ * Copyright (c) 2022 Matthew Nelson
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ **/
+package io.matthewnelson.kmp.tor.controller.common.config
+
+import io.matthewnelson.kmp.tor.common.annotation.InternalTorApi
+import io.matthewnelson.kmp.tor.controller.common.config.TorConfig.Option.*
+import kotlin.test.*
+
+@Suppress("ClassName")
+@OptIn(InternalTorApi::class)
+class TorConfig_Option_Time_UnitTest {
+
+    @Test
+    fun givenMinutes_whenLessThan1_valueIs1Minutes() {
+        assertEquals("1 minutes", Time.Minutes(0).value)
+    }
+
+    @Test
+    fun givenHours_whenLessThan1_valueIs1Hour() {
+        assertEquals("1 hours", Time.Hours(0).value)
+    }
+
+    @Test
+    fun givenDays_whenLessThan1_valueIs1Day() {
+        assertEquals("1 days", Time.Days(0).value)
+    }
+
+    @Test
+    fun givenWeeks_whenLessThan1_valueIs1Week() {
+        assertEquals("1 weeks", Time.Weeks(0).value)
+    }
+}
